Modifying RHEL7 code to use lis clocksource, remove mshyperv.c and hv_timsource.c
