--- hv-rhel7.x/hv/rhel5-hv-driver-install
@@ -1,35 +1,0 @@
-echo "Building Modules"
-
-make -C /lib/modules/$(uname -r)/build M=`pwd` clean
-make -C /lib/modules/$(uname -r)/build M=`pwd` modules
-
-echo "Installing Modules"
-
-cp -f ./*.ko /lib/modules/$(uname -r)/extra/
-
-echo "Generating Module dependencies"
-
-depmod
-
-echo "Installing Modprobe Rules"
-
-
-echo "Saving old initramfs"
-
-if [ ! -e /boot/"initrd-$(uname -r).img-old" ]
-then
-	cp -f /boot/"initrd-$(uname -r).img" /boot/"initrd-$(uname -r).img-old"
-fi
-
-echo "Generating initramfs"
-
-mkinitrd --preload=hv_storvsc "initrd-$(uname -r).img" "$(uname -r)"
-
-echo "Installing new initramfs"
-
-mv -f "initrd-$(uname -r).img" /boot/"initrd-$(uname -r).img"
-
-echo "Config X-Window"
-
-cp -f ./xorg.conf /etc/X11/
-
