+++ hv-rhel7.x/hv/vmbus_drv.c
@@ -45,6 +45,32 @@ static struct tasklet_struct msg_dpc;
 static struct completion probe_event;
 static int irq;
 
+
+int hyperv_panic_event(struct notifier_block *nb,
+                        unsigned long event, void *ptr)
+{
+        struct pt_regs *regs;
+
+        regs = task_pt_regs(current);
+
+        wrmsrl(HV_X64_MSR_CRASH_P0, regs->ip);
+        wrmsrl(HV_X64_MSR_CRASH_P1, regs->ax);
+        wrmsrl(HV_X64_MSR_CRASH_P2, regs->bx);
+        wrmsrl(HV_X64_MSR_CRASH_P3, regs->cx);
+        wrmsrl(HV_X64_MSR_CRASH_P4, regs->dx);
+
+        /*
+         * Let Hyper-V know there is crash data available
+         */
+        wrmsrl(HV_X64_MSR_CRASH_CTL, HV_CRASH_CTL_CRASH_NOTIFY);
+        return NOTIFY_DONE;
+}
+
+static struct notifier_block hyperv_panic_block = {
+        .notifier_call = hyperv_panic_event,
+};
+
+
 struct resource hyperv_mmio = {
 	.name  = "hyperv mmio",
 	.flags = IORESOURCE_MEM,
@@ -737,6 +763,36 @@ static void vmbus_flow_handler(unsigned int irq, struct irq_desc *desc)
 	desc->action->handler(irq, desc->action->dev_id);
 }
 
+#define HV_CLOCK_SHIFT  22
+
+static cycle_t read_hv_clock(struct clocksource *arg)
+{
+        cycle_t current_tick;
+        /*
+         * Read the partition counter to get the current tick count. This count
+         * is set to 0 when the partition is created and is incremented in
+         * 100 nanosecond units.
+         */
+        rdmsrl(HV_X64_MSR_TIME_REF_COUNT, current_tick);
+        return current_tick;
+}
+
+static struct clocksource hyperv_cs = {
+        .name           = "hyperv_clocksource_lis",
+        .rating         = 499, /* use this when running on Hyperv*/
+        .read           = read_hv_clock,
+        .mask           = CLOCKSOURCE_MASK(64),
+        /*
+         * The time ref counter in HyperV is in 100ns units.
+         * The definition of mult is:
+         * mult/2^shift = ns/cyc = 100
+         * mult = (100 << shift)
+         */
+        .mult           = (100 << HV_CLOCK_SHIFT),
+        .shift          = HV_CLOCK_SHIFT,
+        .flags          = CLOCK_SOURCE_VALID_FOR_HRES,
+};
+
 /*
  * vmbus_bus_init -Main vmbus driver initialization routine.
  *
@@ -799,8 +855,17 @@ static int vmbus_bus_init(int irq)
 	if (ret)
 		goto err_alloc;
 
-	vmbus_request_offers();
-
+	/*
+         * Only register if the crash MSRs are available
+         */
+        if (ms_hyperv.features & HV_FEATURE_GUEST_CRASH_MSR_AVAILABLE) {
+                atomic_notifier_chain_register(&panic_notifier_list,
+                                               &hyperv_panic_block);
+        }
+        vmbus_request_offers();
+        if (ms_hyperv.features & HV_X64_MSR_TIME_REF_COUNT_AVAILABLE)
+                clocksource_register(&hyperv_cs);
+	
 	return 0;
 
 err_alloc:
